+++ front/lib/api/poke/plugins/workspaces/revoke_users.ts
@@ -1,7 +1,8 @@
 import { revokeAndTrackMembership } from "@app/lib/api/membership";
 import { createPlugin } from "@app/lib/api/poke/types";
 import { UserResource } from "@app/lib/resources/user_resource";
-import { concurrentExecutor, Err, Ok } from "@app/types";
+import { concurrentExecutor } from "@app/lib/utils/async_utils";
+import { Err, Ok } from "@app/types";
 
 export const revokeUsersPlugin = createPlugin({
   manifest: {
