+++ front/lib/api/tracker.ts
@@ -4,9 +4,10 @@ import config from "@app/lib/api/config";
 import { sendEmailWithTemplate } from "@app/lib/api/email";
 import { Authenticator } from "@app/lib/auth";
 import { TrackerConfigurationResource } from "@app/lib/resources/tracker_resource";
+import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import type { Logger } from "@app/logger/logger";
 import type { TrackerGenerationToProcess } from "@app/types";
-import { concurrentExecutor, CoreAPI, removeNulls } from "@app/types";
+import { CoreAPI, removeNulls } from "@app/types";
 
 const TRACKER_FROM_EMAIL = "dev@dust.tt";
 const TRACKER_FROM_NAME = "Bob Tracker"; // 😬
