+++ front/scripts/workspace_helpers.ts
@@ -1,7 +1,7 @@
 import { Workspace } from "@app/lib/models/workspace";
+import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 import type { LightWorkspaceType } from "@app/types";
-import { concurrentExecutor } from "@app/types";
 
 export async function runOnAllWorkspaces(
   worker: (workspace: LightWorkspaceType) => Promise<void>,
